+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/assets/articulation/articulation.py
@@ -9,6 +9,7 @@
 from __future__ import annotations
 
 import torch
+import warnings
 from collections.abc import Sequence
 from prettytable import PrettyTable
 from typing import TYPE_CHECKING
@@ -50,8 +51,7 @@ class Articulation(RigidObject):
     The articulation class is a subclass of the :class:`RigidObject` class. Therefore, it inherits
     all the functionality of the rigid object class. In case of an articulation, the :attr:`root_physx_view`
     attribute corresponds to the articulation root view and can be used to access the articulation
-    related data. The :attr:`body_physx_view` attribute corresponds to the rigid body view of the articulated
-    links and can be used to access the rigid body related data.
+    related data.
 
     The articulation class also provides the functionality to augment the simulation of an articulated
     system with custom actuator models. These models can either be explicit or implicit, as detailed in
@@ -115,12 +115,12 @@ class Articulation(RigidObject):
     @property
     def num_joints(self) -> int:
         """Number of joints in articulation."""
-        return self.root_physx_view.max_dofs
+        return self.root_physx_view.shared_metatype.dof_count
 
     @property
     def num_bodies(self) -> int:
         """Number of bodies in articulation."""
-        return self.root_physx_view.max_links
+        return self.root_physx_view.shared_metatype.link_count
 
     @property
     def joint_names(self) -> list[str]:
@@ -128,11 +128,36 @@ class Articulation(RigidObject):
         return self.root_physx_view.shared_metatype.dof_names
 
     @property
+    def body_names(self) -> list[str]:
+        """Ordered names of bodies in articulation."""
+        return self.root_physx_view.shared_metatype.link_names
+
+    @property
     def root_physx_view(self) -> physx.ArticulationView:
+        """Articulation view for the asset (PhysX).
+
+        Note:
+            Use this view with caution. It requires handling of tensors in a specific way.
+        """
         return self._root_physx_view
 
     @property
     def body_physx_view(self) -> physx.RigidBodyView:
+        """Rigid body view for the asset (PhysX).
+
+        .. deprecated:: v0.3.0
+
+            In previous versions, this attribute returned the rigid body view over all the links of the articulation.
+            However, this led to confusion with the link ordering as they were not ordered in the same way as the
+            articulation view.
+
+            Therefore, this attribute will be removed in v0.3.0. Please use the :attr:`root_physx_view` attribute
+            instead.
+
+        """
+        dep_msg = "The attribute 'body_physx_view' will be removed in v0.3.0. Please use 'root_physx_view' instead."
+        warnings.warn(dep_msg, DeprecationWarning)
+        carb.log_error(dep_msg)
         return self._body_physx_view
 
     """
@@ -154,28 +179,46 @@ class Articulation(RigidObject):
         If any explicit actuators are present, then the actuator models are used to compute the
         joint commands. Otherwise, the joint commands are directly set into the simulation.
         """
-        super().write_data_to_sim()
-        # write commands
+        # write external wrench
+        if self.has_external_wrench:
+            # apply external forces and torques
+            self._body_physx_view.apply_forces_and_torques_at_position(
+                force_data=self._external_force_body_view_b.view(-1, 3),
+                torque_data=self._external_torque_body_view_b.view(-1, 3),
+                position_data=None,
+                indices=self._ALL_BODY_INDICES,
+                is_global=False,
+            )
+
+        # apply actuator models
         self._apply_actuator_model()
-        # apply actions into simulation
+        # write actions into simulation
         self.root_physx_view.set_dof_actuation_forces(self._joint_effort_target_sim, self._ALL_INDICES)
         # position and velocity targets only for implicit actuators
         if self._has_implicit_actuators:
             self.root_physx_view.set_dof_position_targets(self._joint_pos_target_sim, self._ALL_INDICES)
             self.root_physx_view.set_dof_velocity_targets(self._joint_vel_target_sim, self._ALL_INDICES)
 
-    def update(self, dt: float | None = None):
+    def update(self, dt: float):
         # -- root state (note: we roll the quaternion to match the convention used in Isaac Sim -- wxyz)
         self._data.root_state_w[:, :7] = self.root_physx_view.get_root_transforms()
         self._data.root_state_w[:, 3:7] = math_utils.convert_quat(self._data.root_state_w[:, 3:7], to="wxyz")
         self._data.root_state_w[:, 7:] = self.root_physx_view.get_root_velocities()
+
+        # -- body-state (note: we roll the quaternion to match the convention used in Isaac Sim -- wxyz)
+        self._data.body_state_w[..., :7] = self.root_physx_view.get_link_transforms()
+        self._data.body_state_w[..., 3:7] = math_utils.convert_quat(self._data.body_state_w[..., 3:7], to="wxyz")
+        self._data.body_state_w[..., 7:] = self.root_physx_view.get_link_velocities()
+
         # -- joint states
         self._data.joint_pos[:] = self.root_physx_view.get_dof_positions()
         self._data.joint_vel[:] = self.root_physx_view.get_dof_velocities()
         self._data.joint_acc[:] = (self._data.joint_vel - self._previous_joint_vel) / dt
+
         # -- update common data
         # note: these are computed in the base class
         self._update_common_data(dt)
+
         # -- update history buffers
         self._previous_joint_vel[:] = self._data.joint_vel[:]
 
@@ -199,6 +242,24 @@ class Articulation(RigidObject):
             joint_subset = self.joint_names
         # find joints
         return string_utils.resolve_matching_names(name_keys, joint_subset)
+
+    """
+    Operations - Setters.
+    """
+
+    def set_external_force_and_torque(
+        self,
+        forces: torch.Tensor,
+        torques: torch.Tensor,
+        body_ids: Sequence[int] | slice | None = None,
+        env_ids: Sequence[int] | None = None,
+    ):
+        # call parent to set the external forces and torques into buffers
+        super().set_external_force_and_torque(forces, torques, body_ids, env_ids)
+        # reordering of the external forces and torques to match the body view ordering
+        if self.has_external_wrench:
+            self._external_force_body_view_b = self._external_force_b[:, self._body_view_ordering]
+            self._external_torque_body_view_b = self._external_torque_b[:, self._body_view_ordering]
 
     """
     Operations - Writers.
@@ -500,6 +561,17 @@ class Articulation(RigidObject):
         body_names_regex = r"(" + "|".join(physx_body_names) + r")"
         body_names_regex = f"{self.cfg.prim_path}/{body_names_regex}"
         self._body_physx_view = self._physics_sim_view.create_rigid_body_view(body_names_regex.replace(".*", "*"))
+
+        # create ordering from articulation view to body view for body names
+        # note: we need to do this since the body view is not ordered in the same way as the articulation view
+        # -- root view
+        root_view_body_names = self.body_names
+        # -- body view
+        prim_paths = self._body_physx_view.prim_paths[: self.num_bodies]
+        body_view_body_names = [path.split("/")[-1] for path in prim_paths]
+        # -- mapping from articulation view to body view
+        self._body_view_ordering = [body_view_body_names.index(name) for name in root_view_body_names]
+        self._body_view_ordering = torch.tensor(self._body_view_ordering, dtype=torch.long, device=self.device)
 
         # log information about the articulation
         carb.log_info(f"Articulation initialized at: {self.cfg.prim_path} with root '{root_prim_path_expr}'.")